Rename: school, major repository 패키지 이동
