+++ src/test/java/com/gwakkili/devbe/mail/MailServiceTests.java
@@ -5,7 +5,7 @@ import com.gwakkili.devbe.exception.customExcption.NotFoundException;
 import com.gwakkili.devbe.exception.customExcption.UnsupportedException;
 import com.gwakkili.devbe.mail.service.MailServiceImpl;
 import com.gwakkili.devbe.mail.repository.MailAuthKeyRepository;
-import com.gwakkili.devbe.major.repository.SchoolRepository;
+import com.gwakkili.devbe.shcool.repository.SchoolRepository;
 import jakarta.mail.MessagingException;
 import jakarta.mail.Session;
 import jakarta.mail.internet.MimeMessage;
